Split test cases into separate files
Each file has its own test cases. If needed split them even further.
